requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
bcrypt==4.2.1
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.2
daphne==4.1.2
Django==5.1.6
django-celery-beat==2.8.0
django-cors-headers==4.7.0
django-filter==25.1
django-mail-templated==2.6.5
django-redis==6.0.0
django-timezone-field==7.1
djangorestframework==3.15.2
dnspython==2.7.0
et_xmlfile==2.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.3
Markdown==3.7
msgpack==1.1.0
openpyxl==3.1.5
pillow==11.2.1
prompt_toolkit==3.0.51
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pymongo==4.11.1
pyOpenSSL==25.0.0
PyPDF2==3.0.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
qrcode==8.2
redis==5.2.1
reportlab==4.4.1
requests==2.32.3
service-identity==24.2.0
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
zope.interface==7.2
